exercise/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
django-rest-swagger==2.2.0
django-environ==0.9.0
model-bakery==1.6.0
pytest==7.1.2
pytest-django==4.5.2
django-filter==22.1
pytest-django==4.5.2
psycopg2>=2.8
model-bakery==1.6.0
